EVG-15595: move globals collection and model into model package
