+++ model/global/db.go
@@ -1,25 +1,27 @@
-package db
+package global
 
 import (
+	"github.com/evergreen-ci/evergreen/db"
+	"github.com/mongodb/anser/bsonutil"
 	adb "github.com/mongodb/anser/db"
 	"go.mongodb.org/mongo-driver/bson"
 )
 
 const (
-	GlobalsCollection = "globals"
+	// Collection is the name of the MongoDB collection that stores global
+	// internal tracking information.
+	Collection = "globals"
 )
 
-// Global stores internal tracking information.
-type Global struct {
-	BuildVariant    string `bson:"_id"`
-	LastBuildNumber uint64 `bson:"last_build_number"`
-	LastTaskNumber  uint64 `bson:"last_task_number"`
-}
+var (
+	BuildVariantKey    = bsonutil.MustHaveTag(Global{}, "BuildVariant")
+	LastBuildNumberKey = bsonutil.MustHaveTag(Global{}, "LastBuildNumber")
+)
 
 // GetNewBuildVariantBuildNumber atomically gets a new number for a build,
 // given its variant name.
 func GetNewBuildVariantBuildNumber(buildVariant string) (uint64, error) {
-	session, db, err := GetGlobalSessionFactory().GetSession()
+	session, db, err := db.GetGlobalSessionFactory().GetSession()
 	if err != nil {
 		return 0, err
 	}
@@ -27,18 +29,18 @@ func GetNewBuildVariantBuildNumber(buildVariant string) (uint64, error) {
 
 	// get the record for this build variant
 	global := Global{}
-	err = db.C(GlobalsCollection).Find(bson.M{"_id": buildVariant}).One(&global)
+	err = db.C(Collection).Find(bson.M{BuildVariantKey: buildVariant}).One(&global)
 
 	// if this is the first build for this
 	// this build variant, insert it
 	if err != nil && adb.ResultsNotFound(err) {
 		change := adb.Change{
-			Update:    bson.M{"$inc": bson.M{"last_build_number": 1}},
+			Update:    bson.M{"$inc": bson.M{LastBuildNumberKey: 1}},
 			Upsert:    true,
 			ReturnNew: true,
 		}
 
-		_, err = db.C(GlobalsCollection).Find(bson.M{"_id": buildVariant}).Apply(change, &global)
+		_, err = db.C(Collection).Find(bson.M{BuildVariantKey: buildVariant}).Apply(change, &global)
 		if err != nil {
 			return 0, err
 		}
@@ -49,11 +51,11 @@ func GetNewBuildVariantBuildNumber(buildVariant string) (uint64, error) {
 	// seen this build variant before
 	// find and modify last build variant number
 	change := adb.Change{
-		Update:    bson.M{"$inc": bson.M{"last_build_number": 1}},
+		Update:    bson.M{"$inc": bson.M{LastBuildNumberKey: 1}},
 		ReturnNew: true,
 	}
 
-	_, err = db.C(GlobalsCollection).Find(bson.M{"_id": buildVariant}).Apply(change, &global)
+	_, err = db.C(Collection).Find(bson.M{BuildVariantKey: buildVariant}).Apply(change, &global)
 
 	if err != nil {
 		return 0, err
