+++ model/global/doc.go
@@ -1,0 +1,3 @@
+// Package global contains the model for global build variant tracking
+// information.
+package global
